refactor: move sqlitedb.ts to src-electron/db/sqlite-db.ts, move SQL statements to sql.ts, prepare queries once
